+++ internal/repository/repository.go
@@ -3,8 +3,8 @@ package repository
 import "context"
 
 type SubscriberRepository interface {
-	GetAll(ctx context.Context) ([]int, error)
-	Add(ctx context.Context, id int) error
-	Del(ctx context.Context, id int) error
+	GetAll(ctx context.Context) ([]int64, error)
+	Add(ctx context.Context, id int64) error
+	Del(ctx context.Context, id int64) error
 	Close()
 }
